trip_planner/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0,<4.0
Pillow>=8.0.0,<9.0.0
gunicorn>=20.0.4,<21.0.0
psycopg2-binary>=2.8.6,<3.0.0
django-debug-toolbar>=3.2.0,<4.0.0
requests>=2.25.0,<3.0.0
whitenoise>=5.2.0,<6.0.0
debugpy==1.5.1
dj-database-url
psycopg2-binary

